[FIX] LDAP sync admin action was not syncing exisent users (#16671)
