+++ app/ldap/server/syncUsers.js
@@ -1,6 +1,6 @@
 import { Meteor } from 'meteor/meteor';
 
-import { importNewUsers } from './sync';
+import { sync } from './sync';
 import { hasRole } from '../../authorization';
 import { settings } from '../../settings';
 
@@ -21,7 +21,7 @@ Meteor.methods({
 
 		this.unblock();
 
-		importNewUsers();
+		sync();
 
 		return {
 			message: 'Sync_in_progress',
