+++ server/methods/setUserActiveStatus.js
@@ -1,11 +1,8 @@
 import { Meteor } from 'meteor/meteor';
 import { check } from 'meteor/check';
-import { Accounts } from 'meteor/accounts-base';
 
-import * as Mailer from '../../app/mailer';
 import { hasPermission } from '../../app/authorization';
-import { Users, Subscriptions } from '../../app/models';
-import { settings } from '../../app/settings';
+import { setUserActiveStatus } from '../../app/lib/server/functions/setUserActiveStatus';
 
 Meteor.methods({
 	setUserActiveStatus(userId, active) {
@@ -24,42 +21,7 @@ Meteor.methods({
 			});
 		}
 
-		const user = Users.findOneById(userId);
-
-		if (!user) {
-			return false;
-		}
-
-		Users.setUserActive(userId, active);
-
-		if (user.username) {
-			Subscriptions.setArchivedByUsername(user.username, !active);
-		}
-
-		if (active === false) {
-			Users.unsetLoginTokens(userId);
-		} else {
-			Users.unsetReason(userId);
-		}
-		if (active && !settings.get('Accounts_Send_Email_When_Activating')) {
-			return true;
-		}
-		if (!active && !settings.get('Accounts_Send_Email_When_Deactivating')) {
-			return true;
-		}
-
-		const destinations = Array.isArray(user.emails) && user.emails.map((email) => `${ user.name || user.username }<${ email.address }>`);
-
-
-		const email = {
-			to: destinations,
-			from: settings.get('From_Email'),
-			subject: Accounts.emailTemplates.userActivated.subject({ active }),
-			html: Accounts.emailTemplates.userActivated.html({ active, name: user.name, username: user.username }),
-		};
-
-		Mailer.sendNoWrap(email);
-
+		setUserActiveStatus(userId, active);
 
 		return true;
 	},
